refactor(memory): own dreaming contracts in host sdk
